requirements.txt
streamlit>=1.32.0
pandas>=2.1.0
plotly>=5.18.0
google-genai>=1.0.0
